On edit, trigger label animations
A small update to get the labels to animate when updating.
This is done by adding a conditional class to the labels:
:class=" [ id ? 'active' : '']"

The problem was when clicking update on a past post, it would not trigger a change event on the inputs. This is what caused the labels to not animate up and caused it to sit on top of the input. I tried to trigger a change event but this did not update the labels to have the active class.
Instead if the edit post window has an id, it will have the active class.